refactor(ChatPage): extract logic into hook
